Resolves issues using play_file and start/stop_tone
Added `deinit()` to both. You still get a RuntimeError of `DAC
already in use` if you are currently playing a tone using
`start_tone` and then try to `play_file` before ending the tone.
Added `try` and `except` to `play_file` to avoid this error if
these functions are used in this manner.